shop/serializers/color_serializer.py
from rest_framework import serializers
from shop.models.color import Color


class ColorSerializer(serializers.ModelSerializer):
    class Meta:
        model = Color
        fields = ['id', 'name', 'color']
