requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
autopep8==1.6.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.0.3
django-ckeditor==6.2.0
django-colorfield==0.6.3
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django-js-asset==2.0.0
django-redis==5.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.0
kombu==5.2.4
Markdown==3.3.6
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.0.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyparsing==3.0.7
pytz==2022.1
redis==4.2.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.0.0
wrapt==1.14.0
zipp==3.7.0
